pyproject.toml
[tool.poetry]
name = "industrial-digital-polygon"
version = "0.1.0"
description = "Backend application for industrial digital polygon"
authors = ["Sergey-Gerasimo <gerasimovsergey2004@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "app", from = "."}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["httpx"], version = "^0.110.0"}
uvicorn = {extras = ["all"], version = "^0.28.0"}
ipython = "^8.22.2"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
faker = "^24.3.0"
punq = "^0.7.0"
httpx = "^0.27.0"
pydantic-settings = "^2.2.1"
loguru = "^0.7.3"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.43"}
redis = "^6.4.0"
aio-pika = "^9.5.7"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
asyncpg = "^0.30.0"
aiosqlite = "^0.21.0"
passlib = "^1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
